use WyRand instead of Xoshiro

+++ web3_proxy/src/app/mod.rs
@@ -101,8 +101,8 @@ pub struct AuthorizationChecks {
     pub tracking_level: TrackingLevel,
     /// Chance to save reverting eth_call, eth_estimateGas, and eth_sendRawTransaction to the database.
     /// depending on the caller, errors might be expected. this keeps us from bloating our database
-    /// TODO: f32 would be fine
-    pub log_revert_chance: f64,
+    /// u16::MAX == 100%
+    pub log_revert_chance: u16,
     /// if true, transactions are broadcast only to private mempools.
     /// IMPORTANT! Once confirmed by a miner, they will be public on the blockchain!
     pub private_txs: bool,
